install process work
I have documented the install process up to the point that it works... there are some errors that need to be ironed out the we are having during the installation - There are some work arounds but I want to try to get it working the correct way.  
I implemented a new virtual machine and ran through the install process as I wrote this to make sure that everything up until running `make build` worked correctly. 
I just saw that @Benjmhart advised that we are keeping everything in this repo and I believe there is also a ticket in with IOHK to solve some of these issues. I will get with Ben and see how he wants me to proceed.